+++ spec/inputs/fifo_spec.rb
@@ -6,6 +6,11 @@ describe LogStash::Inputs::Fifo do
 
   let(:tempfile) { Tempfile.new("/tmp/fifo") }
 
+  it "should be non-reloadable" do
+    plugin = LogStash::Plugin.lookup("input", "fifo").new({ "path" => tempfile.path })
+    expect(plugin.reloadable?).to be_falsey
+  end
+
   it "should register without errors" do
     plugin = LogStash::Plugin.lookup("input", "fifo").new({ "path" => tempfile.path })
     expect { plugin.register }.to_not raise_error