+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.9.x
   - Fix events not being processed in a timely manner
+  - Mark plugin as non-reloadable for https://github.com/elastic/logstash/pull/6499
 
 ## 0.9.1
   - Initial version of logstash-input-fifo